sqlx escaping code moved to its natural home, next to the compiler
