+++ sqlx/lexer.ts
@@ -227,132 +227,6 @@ export class SyntaxTreeNode {
   }
 }
 
-export interface ISqlxParseResults {
-  config: string;
-  js: string;
-  sql: string[];
-  incremental: string;
-  preOperations: string[];
-  postOperations: string[];
-  input: { [label: string]: string };
-}
-
-function getEscapedSqlxStatementsForJavaScriptTemplateString(
-  nodes: Array<string | SyntaxTreeNode>
-) {
-  const results = [""];
-  nodes.map(escapeSqlxNodeForJavaScriptTemplateString).forEach(node => {
-    if (typeof node !== "string" && node.type === SyntaxTreeNodeType.SQL_STATEMENT_SEPARATOR) {
-      results.push("");
-      return;
-    }
-    results[results.length - 1] += typeof node === "string" ? node : node.concatenate();
-  });
-  return results;
-}
-
-function escapeSqlxNodeForJavaScriptTemplateString(node: string | SyntaxTreeNode) {
-  if (typeof node === "string") {
-    return node.replace(/\\/g, "\\\\").replace(/\`/g, "\\`");
-  }
-  switch (node.type) {
-    case SyntaxTreeNodeType.SQL_COMMENT:
-      // Any code (i.e. JavaScript placeholder strings) inside comments should not run, so escape it.
-      return node
-        .concatenate()
-        .replace(/`/g, "\\`")
-        .replace(/\${/g, "\\${");
-    case SyntaxTreeNodeType.SQL_LITERAL_STRING:
-      // Literal strings may contain escapes, which we need to double-escape.
-      return node.concatenate().replace(/\\/g, "\\\\");
-  }
-  return node;
-}
-
-export function parseSqlx(code: string): ISqlxParseResults {
-  const results: ISqlxParseResults = {
-    config: "",
-    js: "",
-    sql: [""],
-    incremental: "",
-    preOperations: [""],
-    postOperations: [""],
-    input: {}
-  };
-
-  const rootNode = SyntaxTreeNode.create(code);
-
-  rootNode
-    .children()
-    .filter(SyntaxTreeNode.isSyntaxTreeNode)
-    .filter(node => node.type === SyntaxTreeNodeType.JAVASCRIPT)
-    .forEach(node => {
-      const concatenated = node.concatenate();
-      if (concatenated.startsWith("config")) {
-        results.config = concatenated.slice("config ".length);
-      } else {
-        results.js += concatenated.slice("js {".length, "}".length * -1);
-      }
-    });
-
-  results.sql = getEscapedSqlxStatementsForJavaScriptTemplateString(
-    rootNode
-      .children()
-      .filter(
-        node =>
-          typeof node === "string" ||
-          [
-            SyntaxTreeNodeType.JAVASCRIPT_TEMPLATE_STRING_PLACEHOLDER,
-            SyntaxTreeNodeType.SQL_COMMENT,
-            SyntaxTreeNodeType.SQL_LITERAL_STRING,
-            SyntaxTreeNodeType.SQL_STATEMENT_SEPARATOR
-          ].includes(node.type)
-      )
-  );
-
-  rootNode
-    .children()
-    .filter(SyntaxTreeNode.isSyntaxTreeNode)
-    .filter(node => node.type === SyntaxTreeNodeType.SQL)
-    .forEach(node => {
-      const firstChild = node.children()[0] as string;
-      const lastChild = node.children().slice(-1)[0] as string;
-
-      const sqlCodeBlockWithoutOuterBraces =
-        node.children().length === 1
-          ? new SyntaxTreeNode(SyntaxTreeNodeType.SQL, [
-              firstChild.slice(firstChild.indexOf("{") + 1, firstChild.lastIndexOf("}"))
-            ])
-          : new SyntaxTreeNode(SyntaxTreeNodeType.SQL, [
-              firstChild.slice(firstChild.indexOf("{") + 1),
-              ...node.children().slice(1, -1),
-              lastChild.slice(0, lastChild.lastIndexOf("}"))
-            ]);
-      const statements = getEscapedSqlxStatementsForJavaScriptTemplateString(
-        sqlCodeBlockWithoutOuterBraces.children()
-      );
-
-      if (firstChild.startsWith("incremental_where")) {
-        if (statements.length > 1) {
-          throw new Error(
-            "'incremental_where' code blocks may only contain a single SQL statement."
-          );
-        }
-        results.incremental = statements[0];
-      } else if (firstChild.startsWith("pre_operations")) {
-        results.preOperations = statements;
-      } else if (firstChild.startsWith("post_operations")) {
-        results.postOperations = statements;
-      } else if (firstChild.startsWith("input")) {
-        if (statements.length > 1) {
-          throw new Error("'input' code blocks may only contain a single SQL statement.");
-        }
-        results.input[firstChild.split('"')[1]] = statements[0];
-      }
-    });
-  return results;
-}
-
 function buildSqlxLexer(): { [x: string]: moo.Rules } {
   const sqlLexer: moo.Rules = {};
   sqlLexer[SQL_LEXER_TOKEN_NAMES.START_CONFIG] = {
